+++ src/navi/js.nim
@@ -33,7 +33,9 @@ claimEntry("navi/js")
 export public, asyncjs
 export jsws.WebSocket, jsws.WsMessage, jsws.WsMessageKind,
        jsws.send, jsws.receive, jsws.close, jsws.closeNormal, jsws.closeGoingAway,
-       jsws.closeMessageTooBig, jsws.WsMessageTooLarge
+       jsws.closeMessageTooBig, jsws.WsMessageTooLarge,
+       jsws.WsReader, jsws.WsWriter, jsws.stream, jsws.streamBinary, jsws.each,
+       jsws.readChunk, jsws.drain, jsws.write
 
 type
   NaviContext* = ref object